+++ include/H2DE/H2DE_engine.h
@@ -3,6 +3,7 @@
 
 #include <SDL2/SDL.h>
 #include <cmath>
+#include <filesystem>
 #include <functional>
 #include <iostream>
 #include <thread>
@@ -10,12 +11,17 @@
 #include <windows.h>
 #include "H2DE_utils.h"
 #include "H2DE_asset_loader.h"
+#include "H2DE_window.h"
+#include "H2DE_camera.h"
+class H2DE_Window;
+class H2DE_Camera;
 
 class H2DE_Engine {
 private:
-    SDL_Window* window;
-    SDL_Renderer* renderer;
-    H2DE_EngineData* data;
+    H2DE_Window* window = nullptr;
+    H2DE_Camera* camera = nullptr;
+
+    H2DE_EngineData data;
     unsigned int fps;
     bool isRunning = true;
     int currentFPS = 0;
@@ -26,7 +32,9 @@ private:
     std::unordered_map<std::string, SDL_Texture*> textures;
     std::unordered_map<std::string, Mix_Chunk*> sounds;
 
-    void run();
+    std::function<void(SDL_Event)> handleEvents = NULL;
+    std::function<void()> update = NULL;
+    std::function<void()> render = NULL;
 
     static int countFilesToLoad(const std::filesystem::path& dir);
     void importFiles(const std::filesystem::path& dir);
@@ -36,8 +44,10 @@ private:
     void assetImported();
 
 public:
-    H2DE_Engine(H2DE_EngineData* data);
+    H2DE_Engine(H2DE_EngineData data);
     ~H2DE_Engine();
+
+    friend void H2DE_RunEngine(H2DE_Engine* engine);
 
     static void H2DE_Delay(unsigned int ms, std::function<void()> callback);
 
@@ -50,20 +60,17 @@ public:
     friend void H2DE_PauseSound(H2DE_Engine* engine, int channel);
     friend void H2DE_ResumeSound(H2DE_Engine* engine, int channel);
 
-    friend H2DE_2DAVector H2DE_GetWindowSize(H2DE_Engine* engine);
-    friend H2DE_2DAVector H2DE_GetWindowMinimumSize(H2DE_Engine* engine);
-    friend H2DE_2DAVector H2DE_GetWindowMaximumSize(H2DE_Engine* engine);
+    friend H2DE_Window* H2DE_GetWindow(H2DE_Engine* engine);
     friend int H2DE_GetFps(H2DE_Engine* engine);
     friend int H2DE_GetCurrentFps(H2DE_Engine* engine);
 
-    friend void H2DE_SetTitle(H2DE_Engine* engine, std::string title);
-    friend void H2DE_SetWindowSize(H2DE_Engine* engine, H2DE_2DAVector size);
-    friend void H2DE_SetWindowMinimumSize(H2DE_Engine* engine, H2DE_2DAVector minSize);
-    friend void H2DE_SetWindowMaximumSize(H2DE_Engine* engine, H2DE_2DAVector maxSize);
     friend void H2DE_SetFps(H2DE_Engine* engine, unsigned int fps);
+    friend void H2DE_SetHandleEventCall(H2DE_Engine* engine, std::function<void(SDL_Event)> call);
+    friend void H2DE_SetUpdateCall(H2DE_Engine* engine, std::function<void()> call);
+    friend void H2DE_SetRenderCall(H2DE_Engine* engine, std::function<void()> call);
 };
 
-extern void H2DE_CreateEngine(H2DE_EngineData* data);
+extern H2DE_Engine* H2DE_CreateEngine(H2DE_EngineData data);
 extern void H2DE_DestroyEngine(H2DE_Engine* engine);
 extern void H2DE_SetVolumeSound(int channel, int volume);