+++ src/H2DE_utils.cpp
@@ -1,3 +1,0 @@
-#include "H2DE_utils.h"
-
-